Scheduling bug fixes
