+++ cocogen/frontend/attribute_scheduler/graph.c
@@ -157,23 +157,23 @@ struct GRnode *GRlookup_node(struct GRgraph *graph, node_st *node,
                                      &lookup_node);
 }
 
-struct GRnode_list *GRget_intranode_successors(struct GRgraph *graph,
-                                               struct GRnode *node) {
-    struct GRnode_list *successors = NULL;
-    for (struct GRnode *succ = graph->nodes; succ != NULL; succ = succ->next) {
-        if (succ == node || succ->node != node->node) {
+struct GRnode_list *GRget_inter_node_dependencies(struct GRgraph *graph,
+                                                  struct GRnode *node) {
+    struct GRnode_list *dependencies = NULL;
+    for (struct GRnode *dep = graph->nodes; dep != NULL; dep = dep->next) {
+        if (dep == node || dep->node != node->node) {
             continue;
         }
 
-        if (GRlookup_edge(graph, node, succ)) {
+        if (GRlookup_edge(graph, dep, node)) {
             struct GRnode_list *entry = MEMmalloc(sizeof(struct GRnode_list));
-            entry->node = succ;
-            entry->next = successors;
-            successors = entry;
+            entry->node = dep;
+            entry->next = dependencies;
+            dependencies = entry;
         }
     }
 
-    return successors;
+    return dependencies;
 }
 
 static struct GRedge *add_edge_internal(struct GRgraph *graph,
